Port to namespace `pvlib.bifacial`
