+++ docs/examples/bifacial/plot_irradiance_nonuniformity_loss.py
@@ -26,7 +26,7 @@ Calculate the incident irradiance lost to non-uniformity in a bifacial PV array
 # from the irradiance levels of each cell in a PV module.
 #
 # The function
-# :py:func:`pvlib.pvsystem.nonuniform_irradiance_deline_power_loss` is
+# :py:func:`pvlib.bifacial.nonuniform_irradiance_deline_power_loss` is
 # used to transform the Relative Mean Absolute Difference (RMAD) of the
 # irradiance into a power loss percentage. Down below you will find a
 # numpy-based implementation of the RMAD function.
@@ -45,7 +45,7 @@ import matplotlib.pyplot as plt
 from matplotlib.cm import ScalarMappable
 from matplotlib.colors import Normalize
 
-from pvlib.pvsystem import nonuniform_irradiance_deline_power_loss
+from pvlib.bifacial import nonuniform_irradiance_deline_power_loss
 
 # %%
 # Problem description
@@ -118,7 +118,7 @@ print(rmad_cells == rmad(cells_irrad[:, 0]))
 # ^^^^^^^^^^^^^
 # Calculate the power loss percentage due to the irradiance non-uniformity
 # with the function
-# :py:func:`pvlib.pvsystem.nonuniform_irradiance_deline_power_loss`.
+# :py:func:`pvlib.bifacial.nonuniform_irradiance_deline_power_loss`.
 
 mismatch_loss = nonuniform_irradiance_deline_power_loss(rmad_cells)
